% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{acplot.mcmc.list}
\alias{acplot.mcmc.list}
\title{Plot the autocorrelation of a Bayesian meta-analysis model}
\usage{
\method{acplot}{mcmc.list}(x, P, greek = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"mcmc.list"}}

\item{P}{Optional dataframe describing the parameters to plot and their respective names}

\item{greek}{Logical value indicating whether parameter labels have to be parsed to get Greek letters. Defaults to FALSE.}

\item{\ldots}{Additional arguments which are currently not used}
}
\description{
Function to display autocorrelation of a fitted Bayesian meta-analysis model.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
\keyword{autocorrelation}
\keyword{meta-analysis}
