% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{acplot}
\alias{acplot}
\title{Plot the autocorrelation of a Bayesian meta-analysis model}
\usage{
acplot(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Function to display autocorrelation of a fitted Bayesian meta-analysis model.
}
\details{
This is a generic function.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
