% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred.R
\name{coef.metapred}
\alias{coef.metapred}
\title{Extract the regression coefficients
The \code{coef} function extracts the estimated model coefficients from objects of class \code{"metapred"}.}
\usage{
\method{coef}{metapred}(object, ...)
}
\arguments{
\item{object}{A fitted \code{metapred} object}

\item{\ldots}{Optional arguments (currently not supported).}
}
\description{
Extract the regression coefficients
The \code{coef} function extracts the estimated model coefficients from objects of class \code{"metapred"}.
}
