% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred.R
\name{confint.listofperf}
\alias{confint.listofperf}
\title{#' @export
ci <- function(object, ...)
  UseMethod("ci", object)}
\usage{
\method{confint}{listofperf}(object, parm, level, ...)
}
\description{
#' @export
ci <- function(object, ...)
  UseMethod("ci", object)
}
