% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.r
\name{cor2cov}
\alias{cor2cov}
\title{Convert a correlation matrix into a covariance matrix}
\usage{
cor2cov(sigma, cormat)
}
\arguments{
\item{sigma}{vector of standard deviations. The order of standard deviations should correspond to the column order in 'cormat'.}

\item{cormat}{a symmetric numeric correlation matrix}
}
\value{
The covariance matrix
}
\description{
Convert a correlation matrix into a covariance matrix
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
