% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valmeta.r
\name{density.valmeta}
\alias{density.valmeta}
\title{Plot density of estimated model parameters}
\usage{
\method{density}{valmeta}(x, par, type, ...)
}
\arguments{
\item{x}{An object of class \code{"valmeta"}}

\item{par}{Character string to specify for which parameter a plot should be generated. Choices are \code{'mu'} 
(mean of the random effects model) and \code{'tau'} (standard deviation of the random effects model).}

\item{type}{Character string to specify whether the prior (\code{'prior}) or posterior distribution (\code{'posterior}) 
should be displayed.}

\item{\ldots}{Additional arguments which are passed to \link{forest}.}
}
\value{
An object of class \code{ggplot}
}
\description{
Function to generate plots of the prior and posterior density of a Bayesian meta-analysis.
}
\examples{
\dontrun{
data(EuroSCORE)
fit <- valmeta(cstat=c.index, cstat.se=se.c.index, cstat.cilb=c.index.95CIl,
               cstat.ciub=c.index.95CIu, N=n, O=n.events, 
               data=EuroSCORE, method="BAYES", slab=Study)
density(fit)
density(fit, type = "posterior")
density(fit, par = "tau", type = "prior")

# Meta-analysis of the O:E ratio
EuroSCORE.new <- EuroSCORE
EuroSCORE.new$n[c(1, 2, 5, 10, 20)] <-  NA
pars <- list(hp.tau.dist="dhalft",   # Prior for the between-study standard deviation
             hp.tau.sigma=1.5,       # Standard deviation for 'hp.tau.dist'
             hp.tau.df=3,            # Degrees of freedom for 'hp.tau.dist'
             hp.tau.max=10)          # Maximum value for the between-study standard deviation
fit2 <- valmeta(measure="OE", O=n.events, E=e.events, N=n, data=EuroSCORE.new,
                method="BAYES", slab=Study, pars=pars)
density(fit2, type = "prior")
} 

}
\references{
Debray TPA, Damen JAAG, Snell KIE, Ensor J, Hooft L, Reitsma JB, et al. A guide to systematic review 
and meta-analysis of prediction model performance. \emph{BMJ}. 2017;356:i6460.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
\keyword{density}
\keyword{meta-analysis}
