% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{dplot.mcmc.list}
\alias{dplot.mcmc.list}
\title{Posterior distribution of estimated model parameters}
\usage{
\method{dplot}{mcmc.list}(x, P, plot_type = "dens", ...)
}
\arguments{
\item{x}{An object of class \code{"mcmc.list"}}

\item{P}{Optional dataframe describing the parameters to plot and their respective names}

\item{plot_type}{Optional character string to specify whether a density plot (\code{"dens"}) or 
histogram (\code{"hist"}) should be displayed.}

\item{\ldots}{Additional arguments which are currently not used}
}
\value{
A \code{ggplot} object.
}
\description{
Generate a plot of the posterior distribution
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
\keyword{density}
\keyword{distribution}
\keyword{meta-analysis}
