% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{dplot}
\alias{dplot}
\title{Posterior distribution of estimated model parameters}
\usage{
dplot(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Generate a plot of the posterior distribution
}
\details{
This is a generic function.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
