% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred.R
\name{fitted.metapred}
\alias{fitted.metapred}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{metapred}(
  object,
  select = "cv",
  step = NULL,
  model = NULL,
  as.stratified = TRUE,
  type = "response",
  ...
)
}
\arguments{
\item{object}{object of class metapred}

\item{select}{character. Select fitted values from "cv" (default) or from "global" model.}

\item{step}{character or numeric. Name or number of step to select if \code{select} = "cv". Defaults to best step.}

\item{model}{character or numeric. Name or number of model to select if \code{select} = "cv". Defaults to
best model.}

\item{as.stratified}{logical. \code{select} = "cv" determines whether returned predictions are stratified in a list 
(\code{TRUE}, default) or in their original order (\code{FALSE}).}

\item{type}{character. Type of fitted value.}

\item{...}{For compatibility only.}
}
\description{
Extract the fitted values of a \code{metapred} object. By default returns fitted values of the model in the 
cross-validation procedure.
}
\details{
Function still under development, use with caution.

Only returns type = "response".
}
\author{
Valentijn de Jong
}
