% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{gelmanplot.mcmc.list}
\alias{gelmanplot.mcmc.list}
\title{Gelman-Rubin-Brooks plot}
\usage{
\method{gelmanplot}{mcmc.list}(x, P, greek = FALSE, ...)
}
\arguments{
\item{x}{An mcmc object}

\item{P}{Optional dataframe describing the parameters to plot and their respective names}

\item{greek}{Logical value indicating whether parameter labels have to be parsed to get Greek letters. Defaults to false.}

\item{\ldots}{Additional arguments which are currently not used}
}
\value{
A \code{ggplot} object.

An object of class \code{ggplot}
}
\description{
This plot shows the evolution of Gelman and Rubin's shrink factor as the number of iterations increases. The code is adapted from
the R package coda.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
