% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{gelmanplot}
\alias{gelmanplot}
\title{Gelman-Rubin-Brooks plot}
\usage{
gelmanplot(...)
}
\arguments{
\item{\ldots}{Additional arguments which are currently not used}
}
\value{
A \code{ggplot} object.

An object of class \code{ggplot}
}
\description{
This plot shows the evolution of Gelman and Rubin's shrink factor as the number of iterations increases. The code is adapted from
the R package coda.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
