% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvmeta.r
\name{gelmanplot.uvmeta}
\alias{gelmanplot.uvmeta}
\title{Gelman-Rubin-Brooks plot}
\usage{
\method{gelmanplot}{uvmeta}(x, confidence = 0.95, ...)
}
\arguments{
\item{x}{An mcmc object}

\item{confidence}{The coverage probability of the confidence interval for the potential scale reduction factor}

\item{\ldots}{Additional arguments which are currently not used}
}
\value{
A \code{ggplot} object.

An object of class \code{ggplot}
}
\description{
This plot shows the evolution of Gelman and Rubin's shrink factor as the number of iterations increases. The code is adapted from
the R package coda.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
