% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred.R
\name{gen}
\alias{gen}
\alias{generalizability}
\title{Generalizability estimates}
\usage{
gen(object, ...)
generalizability(object, ...)
}
\arguments{
\item{object}{A model fit object, either a \link{metapred} or \code{subset(metapred)} object.}

\item{...}{By default, the final model is selected. This parameter allows other arguments to be passed 
to \link{subset.metapred} such that the generalizability estimates of other steps/models may be
returned..}
}
\description{
Obtain generalizability estimates from a model fit.
}
\author{
Valentijn de Jong
}
