% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred.R
\name{is_metapred}
\alias{is_metapred}
\title{Check if an object is of class 'metapred'}
\usage{
is_metapred(object)
}
\arguments{
\item{object}{Object to test.}
}
\value{
Logical. TRUE if object inherits from class 'metapred'.
}
\description{
Check if an object is of class 'metapred'
}
\author{
Valentijn de Jong
}
