% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugsmodels.r
\name{is_min_cstat_set}
\alias{is_min_cstat_set}
\title{Check if pars$min.cstat is set and valid}
\usage{
is_min_cstat_set(pars, check_range = TRUE)
}
\arguments{
\item{pars}{A named list of parameters.}

\item{check_range}{Logical. If TRUE (default), also checks that min.cstat is within (0.5, 1).}
}
\value{
TRUE if pars$min.cstat is set and valid; FALSE otherwise.
}
\description{
Check if pars$min.cstat is set and valid
}
