\name{Kertai}
\alias{Kertai}
\docType{data}

\title{Kertai data}

\description{Data frame with diagnostic accuracy data from exercise electrocardiography.}

\usage{data("Kertai")}

\format{One data frame with 4 variables.
  \describe{
  \item{TP}{integer. number of true positives}  
  \item{FN}{integer. number of false negatives}  
  \item{FP}{integer. number of false positives}  
  \item{TN}{integer. number of true negatives}  
  }
}

\details{The \code{Kertai} data set is a meta-analysis of prognostic test studies and comprises 7 studies where the diagnostic test accuracy of exercise electrocardiography for predicting cardiac events in patients undergoing major vascular surgery was measured.}

\source{
Kertai MD, Boersma E, Bax JJ, Heijenbrok-Kal MH, Hunink MGM, L'talien GJ, Roelandt JRTC, van Urk H, Poldermans D. A meta-analysis comparing the prognostic accuracy of six diagnostic tests for predicting perioperative cardiac risk in patients undergoing major vascular surgery. \emph{Heart} 2003; \bold{89}: 1327--1334.

Jackson D, Riley RD, & White IW. Multivariate meta-analysis: Potential and promise. \emph{Statistics in Medicine} 2010; \bold{30}: 2481--2498.
}

\keyword{datasets}
