% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred_measures.R
\name{ma}
\alias{ma}
\title{Random effects meta-analysis}
\usage{
ma(object, method, ...)
}
\arguments{
\item{object}{A model fit object, such as \link{metapred} object.}

\item{method}{Character, method for meta-analysis passed to \link[metamisc]{valmeta} and \link[metamisc]{uvmeta}.
Defaults to "REML".}

\item{...}{Other arguments passed to \link[metamisc]{metapred}, \link[metamisc]{valmeta} and \link[metamisc]{uvmeta}.}
}
\description{
Meta-analysis of the performance or coefficients of a metapred object.
Caution: it is still under development.
}
\details{
Produces different object types depending on input.
}
\author{
Valentijn de Jong
}
