% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred.R
\name{perf}
\alias{perf}
\alias{performance}
\title{Performance estimates}
\usage{
perf(object, ...)
performance(object, ...)
}
\arguments{
\item{object}{A model fit object, either a \link{metapred} or \code{subset(metapred)} object.}

\item{...}{By default, the final model is selected. This parameter allows other arguments to be
passed to \link{subset.metapred} such that the performance estimates of other steps/models may be
returned..}
}
\description{
Obtain performance estimates from a model fit.
}
\author{
Valentijn de Jong
}
