% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riley.r
\name{predict.riley}
\alias{predict.riley}
\title{Prediction Interval}
\usage{
\method{predict}{riley}(object, ...)
}
\arguments{
\item{object}{A \code{riley} object.}

\item{\dots}{Additional arguments (currently ignored)}
}
\value{
Data frame containing prediction intervals with the summary estimates \code{beta1} and \code{beta2} 
(for effect size data), or with the mean sensitivity and false positive rate (for diagnostic test accuracy data).
}
\description{
Calculates a prediction interval for the summary parameters of Riley's alternative model for bivariate random-effects 
meta-analysis. This interval predicts in what range future observations will fall given what has already been observed.
}
\details{
Prediction intervals are based on Student's t-distribution with (numstudies - 5) degrees of freedom. The width of 
the interval is specified by the significance level chosen during meta-analysis.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
