% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{rmplot.mcmc.list}
\alias{rmplot.mcmc.list}
\title{Plot the running means of a Bayesian meta-analysis model}
\usage{
\method{rmplot}{mcmc.list}(x, P, greek = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{"mcmc.list"}}

\item{P}{Optional dataframe describing the parameters to plot and their respective names}

\item{greek}{Logical value indicating whether parameter labels have to be parsed to get Greek letters. Defaults to FALSE.}

\item{\ldots}{Additional arguments which are passed to ggs_running}
}
\value{
A \code{ggplot} object.
}
\description{
Function to display running means of a fitted Bayesian meta-analysis model.
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
