% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valmeta.r
\name{rmplot.valmeta}
\alias{rmplot.valmeta}
\title{Plot the running means of a Bayesian meta-analysis model}
\usage{
\method{rmplot}{valmeta}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"valmeta"}}

\item{\ldots}{Additional arguments which are currently not used}
}
\value{
A \code{ggplot} object.

An object of class \code{ggplot}
}
\description{
Function to display running means of a fitted Bayesian meta-analysis model.
}
\details{
Results are displayed for the estimated mean (\code{mu}) and standard-deviation (\code{tau}) of the meta-analysis model.
}
\examples{
\dontrun{
data(EuroSCORE)

fit <- valmeta(cstat=c.index, cstat.se=se.c.index, cstat.cilb=c.index.95CIl,
               cstat.ciub=c.index.95CIu, N=n, O=n.events, 
               data=EuroSCORE, method="BAYES", slab=Study)
rmplot(fit)
} 

}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
\keyword{convergence}
\keyword{meta-analysis}
