\name{Roberts}
\alias{Roberts}
\docType{data}

\title{Roberts data}

\description{Data frame with summary data from 14 comparative studies.}

\usage{data("Roberts")}

\format{One data frame with 2 variables.
  \describe{
  \item{SDM}{Effect sizes (standardized differences in means)}  
  \item{SE}{Standard error of the effect sizes}  
  }
}

\details{The \code{Roberts} data set is a meta-analysis of 14 studies comparing 'set shifting' ability (the ability to move back and forth between different tasks) in people with eating disorders and healthy controls.}

\source{
Roberts ME, Tchanturia K, Stahl D, Southgate L, Treasure J. A systematic review and meta-analysis of set-shifting ability in eating disorders. \emph{Psychological Medicine} 2007, \bold{37}: 1075--1084.

Higgins JPT, Thompson SG, Spiegelhalter DJ. A re-evaluation of random-effects meta-analysis. \emph{Journal of the Royal Statistical Society. Series A (Statistics in Society)} 2009, \bold{172}: 137--159.
}

\keyword{datasets}
