% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccalc.r
\name{ccalc}
\alias{ccalc}
\title{Calculate the concordance statistic}
\usage{
ccalc(cstat, cstat.se, cstat.cilb, cstat.ciub, cstat.cilv, sd.LP, N, O, Po,
  data, slab, g = NULL, pars, ...)
}
\arguments{
\item{cstat}{vector to specify the estimated c-statistics.}

\item{cstat.se}{vector to specify the corresponding standard errors}

\item{cstat.cilb}{vector to specify the lower limits of the confidence interval.}

\item{cstat.ciub}{vector to specify the upper limits of the confidence interval.}

\item{cstat.cilv}{vector to specify the levels of aformentioned confidence interval limits. 
(default: 0.95, which corresponds to the 95\% confidence interval).}

\item{sd.LP}{vector to specify the standard deviations of the linear predictor (prognostic index)}

\item{N}{vector to specify the validation study sizes.}

\item{O}{vector to specify the total number of observed events.}

\item{Po}{vector to specify the observed event probabilities.}

\item{data}{optional data frame containing the variables given to the arguments above.}

\item{slab}{optional vector with labels for the studies.}

\item{g}{a quoted string that is the function to transform estimates of the c-statistic; see the details below.}

\item{pars}{optional list to specify additional arguments: 
significance level of the confidence interval (default: \code{level=0.95}), 
method for calculating the standard error of the c-statistic 
(default: \code{method.restore.c.se="Newcombe.4"}).}

\item{\ldots}{Additional arguments.}
}
\value{
An array with the following columns:
\describe{
 \item{"theta"}{The (transformed) c-statistics. }
 \item{"theta.se"}{Standard errors of the (transformed) c-statistics.}
 \item{"theta.CIl"}{Lower confidence interval of the (transformed) c-statistics. The level is specified in
 \code{pars$level}. Intervals are calculated on the same scale as \code{theta} by assuming a Normal distribution.}
 \item{"theta.CIu"}{Upper confidence interval of the (transformed) c-statistics. The level is specified in
 \code{pars$level}. Intervals are calculated on the same scale as \code{theta} by assuming a Normal distribution.}
 \item{"theta.source"}{Method used for calculating the (transformed) c-statistic.}
 \item{"theta.se.source"}{Method used for calculating the standard error of the (transformed) c-statistic.}
}
}
\description{
The function calculates the (logit transformed) concordance (c-) statistic with the corresponding sampling variance.
}
\details{
The c-statistic is a measure of discrimination, and indicates the ability of a prediction model to 
distinguish between patients developing and not developing the outcome. The c-statistic typically ranges 
from 0.5 (no discriminative ability) to 1 (perfect discriminative ability). 

When performing a meta-analysis of the c-statistic, it is generally recommended to apply the logit transformation 
(Debray et al., 2017; Snell et al., 2017). This can be achieved by specifying \code{g="log(cstat/(1-cstat))"}. Other
transformations can be specified in a similar manner. 

\subsection{Restoring the c-statistic}{
For studies where the c-statistic is missing, it is estimated from the standard deviation of the linear predictor 
(\code{theta.source="std.dev(LP)"). The corresponding method is described by White et al. (2015). }.
}

\subsection{Restoring the standard error of the c-statistic}{
When missing, the standard error of the c-statistic can be estimated from the confidence interval. Alternatively, 
the standard error can be approximated from a combination of the reported c-statistic, the total sample size and 
the total number of events (Debray et al. 2017). This can be achieved by adopting (a modification of) the method 
proposed by Hanley and McNeil, as specified in \code{pars$method.restore.c.se}. Possible options for this argument are
\code{"Newcombe.2"} and \code{"Newcombe.4"}, which correspond to method 2 and, respectively, method 4 in Newcombe (2006).
}
}
\examples{
######### Validation of prediction models with a binary outcome #########
data(EuroSCORE)

# Calculate the c-statistic and its standard error
ccalc(cstat=c.index, cstat.se=se.c.index, cstat.cilb=c.index.95CIl, cstat.ciub=c.index.95CIu, 
      N=n, O=n.events, data=EuroSCORE, slab=Study)
  
# Calculate the logit c-statistic and its standard error
ccalc(cstat=c.index, cstat.se=se.c.index, cstat.cilb=c.index.95CIl, cstat.ciub=c.index.95CIu, 
      N=n, O=n.events, data=EuroSCORE, slab=Study, g="log(cstat/(1-cstat))")
                                                            
}
\references{
\itemize{
\item Debray TPA, Damen JAAG, Snell KIE, Ensor J, Hooft L, Reitsma JB, et al. A guide to systematic review 
and meta-analysis of prediction model performance. \emph{BMJ}. 2017; 356:i6460.
\item Hanley JA, McNeil BJ. The meaning and use of the area under a receiver operating characteristic (ROC) 
curve. \emph{Radiology}. 1982; 143(1):29--36.
\item Newcombe RG. Confidence intervals for an effect size measure based on the Mann-Whitney statistic. 
Part 2: asymptotic methods and evaluation. \emph{Stat Med}. 2006; 25(4):559--73.
\item Snell KI, Ensor J, Debray TP, Moons KG, Riley RD. Meta-analysis of prediction model performance across 
multiple studies: Which scale helps ensure between-study normality for the C -statistic and calibration measures? 
\emph{Statistical Methods in Medical Research}. 2017. 
\item White IR, Rapsomaniki E, the Emerging Risk Factors Collaboration. Covariate-adjusted measures of discrimination 
for survival data. \emph{Biom J}. 2015;57(4):592--613. 
}
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
\keyword{concordance}
\keyword{discrimination}
\keyword{meta-analysis}
\keyword{performance}
\keyword{statistic}
