% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccalc.r
\name{ccalc}
\alias{ccalc}
\title{Calculate the concordance statistic}
\usage{
ccalc(cstat, cstat.se, cstat.95CI, sd.LP, N, O, E, Po, Pe, slab, pars, ...)
}
\arguments{
\item{cstat}{Optional vector with the estimated c-statistic for each valiation}

\item{cstat.se}{Optional vector with the standard error of the estimated c-statistics}

\item{cstat.95CI}{Optional 2-dimensional array with the lower (first column) and upper (second column) boundary 
of the 95\% confidence interval of the estimated c-statistics}

\item{sd.LP}{Optional vector with the standard deviation of the linear predictor (prognostic index) for each validation}

\item{N}{Optional vector with the total number of participants for each valiation}

\item{O}{Optional vector with the total number of observed events for each valiation
(if specified, during time \code{t.val})}

\item{E}{Optional vector with the total number of expected events for each valiation 
(if specified, during time \code{t.val})}

\item{Po}{Optional vector with the (cumulative) observed event probability for each valiation}

\item{Pe}{Optional vector with the (cumulative) expected event probability for each validation}

\item{slab}{Optional vector with study names}

\item{pars}{A list with additional arguments: significance level of the confidence intervals (default: \code{level=0.95}), 
method for calculating the standard error of the c-statistic 
(default: \code{method.restore.c.se="Newcombe.4"}), and transformation for the c-statistic
(default: \code{model = "normal/logit"}, which corresponds to the logit transformation). see "Details" for more information.}

\item{\ldots}{Additional arguments.}
}
\value{
An array with the following columns:
\describe{
 \item{"theta"}{The (transformed) c-statistics. }
 \item{"theta.se"}{Standard errors of the (transformed) c-statistics.}
 \item{"theta.CIl"}{Lower confidence interval of the (transformed) c-statistics. The level is specified in
 \code{pars$level}. Intervals are calculated on the same scale as \code{theta} by assuming a Normal distribution.}
 \item{"theta.CIu"}{Upper confidence interval of the (transformed) c-statistics. The level is specified in
 \code{pars$level}. Intervals are calculated on the same scale as \code{theta} by assuming a Normal distribution.}
 \item{"theta.source"}{Method used for calculating the (transformed) c-statistic.}
 \item{"theta.se.source"}{Method used for calculating the standard error of the (transformed) c-statistic.}
}
}
\description{
The function calculates the (logit transformed) concordance (c-) statistic with the corresponding sampling variance.
}
\details{
The c-statistic is a measure of discrimination, and indicates the ability of a prediction model to 
distinguish between patients developing and not developing the outcome. The c-statistic typically ranges 
from 0.5 (no discriminative ability) to 1 (perfect discriminative ability). 

By default, the logit transformation is applied to the c-statistic. If no transformation is needed, 
simply set \code{pars$model} equal to \code{"normal/identity"}.

\subsection{Restoring the c-statistic}{
For studies where the c-statistic is missing, it is estimated from the standard deviation of the linear predictor 
(\code{theta.source="std.dev(LP)"). The corresponding method is described by White et al. (2015). }.
}

\subsection{Restoring the standard error of the c-statistic}{
When missing, the standard error of the c-statistic can be estimated from the confidence interval. Alternatively, 
the standard error can be approximated from a combination of the reported c-statistic, the total sample size and 
the total number of events (Debray et al. 2017). This can be achieved by adopting (a modification of) the method 
proposed by Hanley and McNeil, as specified in \code{pars$method.restore.c.se}. Possible options for this argument are
\code{"Newcombe.2"} and \code{"Newcombe.4"}, which correspond to method 2 and, respectively, method 4 in Newcombe (2006).
}
}
\examples{
######### Validation of prediction models with a binary outcome #########
data(EuroSCORE)

# Calculate the logit c-statistic and its standard error
with(EuroSCORE, ccalc(cstat=c.index, cstat.se=se.c.index, 
                      cstat.95CI=cbind(c.index.95CIl,c.index.95CIu), 
                      N=n, O=n.events, slab=Study))
  
# Calculate the c-statistic and its standard error
with(EuroSCORE, ccalc(cstat=c.index, cstat.se=se.c.index, 
                      cstat.95CI=cbind(c.index.95CIl,c.index.95CIu), 
                      N=n, O=n.events, slab=Study, pars=list(model="normal/identity")))
                                                            
}
\references{
\itemize{
\item Debray TPA, Damen JAAG, Snell KIE, Ensor J, Hooft L, Reitsma JB, et al. A guide to systematic review 
and meta-analysis of prediction model performance. \emph{BMJ}. 2017; 356:i6460.
\item Hanley JA, McNeil BJ. The meaning and use of the area under a receiver operating characteristic (ROC) 
curve. \emph{Radiology}. 1982; 143(1):29--36.
\item Newcombe RG. Confidence intervals for an effect size measure based on the Mann-Whitney statistic. 
Part 2: asymptotic methods and evaluation. \emph{Stat Med}. 2006; 25(4):559--73.
\item White IR, Rapsomaniki E, the Emerging Risk Factors Collaboration. Covariate-adjusted measures of discrimination 
for survival data. \emph{Biom J}. 2015;57(4):592--613. 
}
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
\keyword{calibration}
\keyword{discrimination}
\keyword{extraction}
\keyword{meta-analysis}
