% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fat.r
\name{fat}
\alias{fat}
\title{Regression tests for detecting funnel plot asymmetry}
\usage{
fat(b, b.se, method = "E-UW")
}
\arguments{
\item{b}{Vector with the effect size of each study. Examples are log odds ratio, log hazards ratio, log relative risk.}

\item{b.se}{vector with the standard error of the effect size of each study}

\item{method}{Method for testing funnel plot asymmetry, defaults to \code{"E-UW"} (Egger's test). More info in "Details".}
}
\value{
a list containing the following entries:
\itemize{
 \item{"pval"} {P-value indicating statistical significance of funnel plot asymettry test}
 \item{"model"}{A fitted \code{glm} object, reprsenting the estimated regression model}
}
}
\description{
Formal tests to assess the presence of funnel plot asymmetry typically estimate the association between the reported effect size and their standard error, the total sample size or the inverse of the total sample size
}
\details{
A common method to test the presence of small-study effects (\code{method="E-UW"}) is given as the following unweighted regression model: \deqn{\hat{b}_k = \beta_0 + \beta_1\, \widehat \mathrm{SE}(\hat{b}_k) + \epsilon_k \;,\; \epsilon_k \sim \mathcal{N}\left(0, \sigma^2 \right) }{b=A+B*b.se + e;  e[~N(0,s^2)}
Whereas \eqn{\beta_0}{A} indicates the size and direction of the treatment effect, \eqn{\beta_1}{B} provides a measure of asymmetry; the larger its deviation from zero the more pronounced the asymmetry.
}
\examples{
data(Fibrinogen)
b <- log(Fibrinogen$HR)
b.se <- ((log(Fibrinogen$HR.975) - log(Fibrinogen$HR.025))/(2*qnorm(0.975)))
fat(b=b, b.se=b.se)

}
\references{
Debray TPA, Moons KGM, Riley RD. Detecting small-study effects and funnel plot asymmetry in meta-analysis of survival data: a comparison of new and existing tests. \emph{Res Syn Meth}. 2017 Oct 3.
}
\author{
Thomas Debray
}
