% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred_measures.R
\name{forest.metapred}
\alias{forest.metapred}
\title{Forest plot of a metapred fit}
\usage{
\method{forest}{metapred}(object, step = NULL, model = NULL, ...)
}
\arguments{
\item{object}{A \code{metapred} fit object.}

\item{step}{Which step should be plotted? Defaults to the best step. numeric is converted to name of the step: 0 for 
an unchanged model, 1 for the first change...}

\item{model}{Which model change should be plotted? NULL (default, best change) or character name of variable or (integer) 
index of model change.}

\item{...}{For compatibility only.}
}
\description{
Draw a forest plot of the performance of an internally-externally cross-validated model. By default the final model is shown.
}
\author{
Valentijn de Jong <Valentijn.M.T.de.Jong@gmail.com>

Valentijn de Jong <Valentijn.M.T.de.Jong@gmail.com>
}
