% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred_measures.R
\name{forest.mp.cv.val}
\alias{forest.mp.cv.val}
\alias{forest.perf}
\title{Forest plot of a validation object.}
\usage{
\method{forest}{mp.cv.val}(object, ...)
}
\arguments{
\item{object}{An \code{mp.cv.val} or \code{perf} object.}

\item{...}{For compatibility only.}
}
\description{
Draw a forest plot of the performance of an internally-externally cross-validated model.
}
\author{
Valentijn de Jong <Valentijn.M.T.de.Jong@gmail.com>

Valentijn de Jong <Valentijn.M.T.de.Jong@gmail.com>
}
