% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.r
\name{forest}
\alias{forest}
\title{Forest plot}
\usage{
forest(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\description{
Generate a forest plot by specifying the various effect sizes, confidence intervals and summary estimate.
}
\details{
This is a generic function. See \link{forest.default} for making forest plots of summary statistics,
\link{forest.metapred} for plotting \link{metapred} objects, and \link{forest.mp.cv.val} for plotting 
\code{mp.cv.val} objects.
}
\author{
Thomas Debray <thomas.debray@gmail.com>

Valentijn de Jong <Valentijn.M.T.de.Jong@gmail.com>
}
