\name{metamisc-package}
\alias{metamisc-package}
\docType{package}
\title{Diagnostic and Prognostic Meta-Analysis}
\description{Meta-analysis of diagnostic and prognostic modeling studies. Summarize estimates of diagnostic test accuracy and prediction model performance. Validate, update and combine published prediction models.}

\details{
%\tabular{ll}{
%Package: \tab metamisc\cr
%Type: \tab Package\cr
%Version: \tab 0.1.3\cr
%Date: \tab 2017-03-09\cr
%License: \tab GPL-2 \cr
%}
The following functionality is currently implemented: univariate meta-analysis of summary data (\code{\link{uvmeta}}), bivariate meta-analysis of correlated outcomes (\code{\link{riley}}), bivariate meta-analysis of diagnostic test accuracy data, meta-analysis of predition model performance (\code{\link{valmeta}}).}

\author{Thomas Debray <thomas.debray@gmail.com>}

\references{
Debray TPA, Damen JAAG, Snell KIE, Ensor J, Hooft L, Reitsma JB, et al. A guide to systematic review and meta-analysis of prediction model performance. \emph{BMJ}. 2017; \bold{356}:i6460. 

Riley RD, Thompson JR, Abrams KR. An alternative model for bivariate random-effects meta-analysis when the within-study correlations are unknown. \emph{Biostatistics} 2008; \bold{9}: 172--186.
}

\keyword{package}
\keyword{regression}
\keyword{meta-analysis}
\keyword{validation}

