\name{riley-package}
\alias{riley-package}
\alias{riley}
\docType{package}
\title{
 Diagnostic and prognostic meta-analysis
}
\description{
This package provides functions for diagnostic and prognostic meta-analyses. It estimates univariate, bivariate and multivariate models, and allows the aggregation of previously published prediction models with new data.
}
\details{
\tabular{ll}{
Package: \tab metamisc\cr
Type: \tab Package\cr
Version: \tab 0.0.2\cr
Date: \tab 2012-08-30\cr
License: \tab GPL-2 \cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Thomas Debray <thomas.debray@gmail.com>

Maintainer: Thomas Debray <thomas.debray@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
