% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fat.r
\name{plot.fat}
\alias{plot.fat}
\title{Display results from the funnel plot asymmetry test}
\usage{
\method{plot}{fat}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fat}}

\item{...}{Additional arguments for \code{\link{plot}}. The argument \code{funnel.xlab} can be used to define the x-label axis.
The argument \code{ref} represents a numeric value indicating the fixed or random effects summary estimate. If no value is provided
then it will be retrieved from a fixed effects meta-analysis (if possible). The argument \code{x.rescale} is a character string indicating how effect sizes should be rescaled. Options are \code{exp} 
(e.g. for log odds or log hazard ratios).}
}
\description{
Generates a funnel plot for a fitted \code{fat} object, with boundaries for the 90\% confidence interval.
}
\examples{
data(Fibrinogen)
b <- log(Fibrinogen$HR)
b.se <- ((log(Fibrinogen$HR.975) - log(Fibrinogen$HR.025))/(2*qnorm(0.975)))
n.total <- Fibrinogen$N.total

plot(fat(b=b, b.se=b.se))
plot(fat(b=b, b.se=b.se, n.total=n.total, method="M-FIV"), funnel.xlab="Log hazard ratio")
plot(fat(b=b, b.se=b.se), funnel.xlab="Hazard ratio", x.rescale="exp")
}
