% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred_recalibrate.R
\name{recalibrate}
\alias{recalibrate}
\title{Recalibrate a Prediction Model}
\usage{
recalibrate(object, newdata, intercept = TRUE, estFUN = NULL, ...)
}
\arguments{
\item{object}{A model fit object to be recalibrated, of class \code{metapred, glm} or \code{lm}, and more.}

\item{newdata}{data.frame containing new data set for updating.}

\item{intercept}{logical. Should the intercept be updated?}

\item{estFUN}{Function for model estimation. If left \code{NULL}, the function is automatically retrieved
for \code{metapred} objects. For other objects, the function with name corresponding to the first class
of the object is taken. E.g. \code{glm()} for \code{glm} objects.}

\item{...}{Optional arguments to pass to \code{estFUN}.}
}
\value{
Updated model fit object, of the same class as \code{object}. Generally, updated coefficients can
be retrieved with \code{coef()}.
}
\description{
\code{recalibrate} is used to recalibrate a prediction model of classes \code{metapred, glm} or  \code{lm},
and more.
}
\details{
Currently only the intercept is updated and leaves other coefficients as well as the variances
and other aspects untouched. For updating the entire model and all its statistics, see \link[stats]{update}.
\code{recalibrate} assumes coefficients are stored in \code{object$coefficients}
and that \code{estFUN} accepts an \code{offset} argument.
}
