% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metapred.R
\name{se}
\alias{se}
\alias{variances}
\title{Standard errors and variances}
\usage{
se(object, ...)
variances(object, ...)
}
\arguments{
\item{object}{A model fit object}

\item{...}{other arguments}
}
\value{
For \code{se} the standard errors of \code{object}, and for 
\code{variances} the variances.
}
\description{
Obtain standard errors or variances of a model fit
}
\author{
Valentijn de Jong
}
