% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvmeta.r
\name{summary.uvmeta}
\alias{summary.uvmeta}
\title{Summarizing Univariate Meta-Analysis Models}
\usage{
\method{summary}{uvmeta}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"uvmeta"}}

\item{\dots}{Optional arguments to be passed on to other functions}
}
\description{
This function provides summary estimates of a fitted univariate meta-analysis model.
}
\references{
\itemize{
\item Borenstein M, Hedges LV, Higgins JPT, Rothstein HR. A basic introduction to fixed-effect and random-effects models for meta-analysis. \emph{Research Synthesis Methods} 2010; \bold{1}: 97--111.
\item DerSimonian R, Laird N. Meta-analysis in clinical trials. \emph{Controlled Clinical Trials} 1986; \bold{7}: 177--188.
\item Riley RD, Higgins JPT, Deeks JJ. Interpretation of random effects meta-analyses. \emph{British Medical Journal} 2011; \bold{342}: d549.
}
}
\seealso{
\code{\link{uvmeta}}
}
\author{
Thomas Debray <thomas.debray@gmail.com>
}
\keyword{DerSimonian}
\keyword{Laird}
\keyword{meta-analysis}
\keyword{random-effects}
\keyword{univariate}
